requirements.txt
langchain==0.3.27
langchain-community==0.3.29
langchain-groq==0.2.4
langchain-openai==0.3.33
langgraph==0.6.6
langchain-text-splitters==0.3.11
rank-bm25==0.2.2
pydantic==2.11.9
python-dotenv==1.0.1
